WordAnalyzer Project
Firstly I create string variable and assign text from chitanka.info. Then I create words array that splits the text by char symbols. Including all the symbols, using removeEmptyEntries to remove all the space that is left. For better understanding I can structure the text with spaces, but they will be gone after I start the program. I create new list, the looped through the words array and if the length of the words is greater than or equal to 3 I add the word to the new list I created. Then I create an integer variable that I assign the count of the words in the list. Now I create method that is used to find the longest word in the text. It's string methods that loops with foreach loop through the words in the array. I assign  to null the new string variable longest. Then I check with simple if check the longest word using the operator or ||. if the check is successful I assign the longest variable to the word that is in the list. At the end of the method I return the longest. In new string variable I assign the method FindLongestWord with the name of the list inside. At the end the program prints the result of the longest word. Now working on the shortest and on the others methods.