ShortestWord Method,added , AverageWord Method added.
Shortest is the same like longest but the if checks is if the word.length is smaller that shortest.length.
AverageLength Method is a little different because is double .
In I create a variable double that is assign to zero.
Then loop through foreach loop and in this loop I assign the variable to length of the words.
Then If check that checks the count and returns the variable divided by the numbers of words. in the else statement returns zero.
Then I print it to the console