Mocking
Installed FakeItEasy package. Created one interface with boolean method. After that, I created DnsService that inherits from the interface, the only thing that method is doing, is returning bool true. In the NetworkService, I Added constructor with parameter IDns. In NetworkServiceTests, I used dependency injection, and FakeItEasy , to fake the IDns, then I added the variable to the NetworkService. In the tests "SendPing", I Call the _dns.SendDNS(),Returns(true). In that way the boolean that will be returned, it will be true. FakeItEasy fakes it and now the if check is going to be passed and the test will work.